backend/requirement.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.20
router==0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
